Add optional lua modules that can operate on given URLs to do a few transformations. The idea here is to externalize those from the parser. The browser emulation will becore a module now. I guess lots of people will also have some neat ideas.
I also seperated the datadir handling since this is not just for TLDs, but also for modules.

I do not want to make faup a bloatware, the intent is still to have the fastest url parser able to handle TLDs retrieval etc. so externalizing extra functions in lua modules as a C library (so you do not have the deal with this if you just use the C API, those are transparent) will help to move it in a wanted direction, with light changes in the parser.

If you do no want to have lua modules, you still have the parser working, but loose some features. This is fine.