Silence a warning about dirent->d_name
d_name is an array so can never be NULL (assuming POSIX compliance). I assume
the test was supposed to make sure that it was safe to test d_name[0] but I
have converted it to make sure there are no entries without file names. Not
sure if that is possible on any OS...